Finish implementation. Task #79 - Have templates based on project type
