+++ VM-Core/src/main/java/com/validation/manager/core/db/Template.java
@@ -24,6 +24,8 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import javax.persistence.NamedQueries;
 import javax.persistence.NamedQuery;
 import javax.persistence.OneToMany;
@@ -70,6 +72,9 @@ public class Template implements Serializable {
     private String templateName;
     @OneToMany(cascade = CascadeType.ALL, mappedBy = "template")
     private List<TemplateNode> templateNodeList;
+    @JoinColumn(name = "project_type_id", referencedColumnName = "id")
+    @ManyToOne(optional = false)
+    private ProjectType projectTypeId;
 
     public Template() {
     }
@@ -107,6 +112,14 @@ public class Template implements Serializable {
 
     public void setTemplateNodeList(List<TemplateNode> templateNodeList) {
         this.templateNodeList = templateNodeList;
+    }
+
+    public ProjectType getProjectTypeId() {
+        return projectTypeId;
+    }
+
+    public void setProjectTypeId(ProjectType projectTypeId) {
+        this.projectTypeId = projectTypeId;
     }
 
     @Override
