--- VM-Core/src/main/resources/db/migration/V1_7__Project_Types.sql
@@ -1,3 +1,0 @@
-INSERT INTO `project_type` (`id`, `type_name`) VALUES (1, 'general.software');
-INSERT INTO `project_type` (`id`, `type_name`) VALUES (2, 'general.hardware');
-INSERT INTO `project_type` (`id`, `type_name`) VALUES (3, 'general.mixed');
